+++ build.gradle
@@ -3,7 +3,7 @@ import static org.gradle.api.JavaVersion.VERSION_1_8
 buildscript {
     ext {
         corda_release_group = 'net.corda'
-        corda_release_version = '4.4'
+        corda_release_version = '4.10'
         tokens_release_group = getProperty("group")
         tokens_release_version = getProperty("version")
         tokens_release_suffix= "$versionSuffix"
