+++ workflows/src/test/kotlin/com/r3/corda/lib/tokens/workflows/LedgerTestWithPersistence.kt
@@ -3,11 +3,11 @@ package com.r3.corda.lib.tokens.workflows
 import com.nhaarman.mockito_kotlin.whenever
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.node.services.IdentityService
+import net.corda.coretesting.internal.rigorousMock
 import net.corda.testing.common.internal.testNetworkParameters
 import net.corda.testing.core.DUMMY_NOTARY_NAME
 import net.corda.testing.core.SerializationEnvironmentRule
 import net.corda.testing.core.TestIdentity
-import net.corda.testing.internal.rigorousMock
 import net.corda.testing.node.MockServices
 import org.junit.Rule
 import org.mockito.Mockito
